Add support for indexing mechabugs

+++ src/java/com/indeed/jiraactions/TsvFileWriter.java
@@ -4,7 +4,9 @@ import com.google.common.base.Joiner;
 import com.indeed.jiraactions.api.customfields.CustomFieldDefinition;
 import com.indeed.jiraactions.api.customfields.CustomFieldValue;
 import com.indeed.util.logging.Loggers;
+
 import org.apache.commons.codec.binary.Base64;
+import org.apache.commons.lang.StringUtils;
 import org.apache.http.HttpResponse;
 import org.apache.http.client.methods.HttpPost;
 import org.apache.http.entity.ContentType;
@@ -59,7 +61,11 @@ public class TsvFileWriter {
     private void createFileAndWriteHeaders(final DateTime day) throws IOException {
         final String filename = String.format("%s_%s.tsv", config.getIndexName(), reformatDate(day));
         final File file = new File(filename);
-        file.deleteOnExit();
+        if (StringUtils.isNotEmpty(config.getIuploadURL())) {
+            file.deleteOnExit();
+        } else {
+            log.info("Not deleting tsv file because upload url is unset.");
+        }
 
         final BufferedWriter bw = new BufferedWriter(new FileWriter(file));
 
@@ -163,6 +169,11 @@ public class TsvFileWriter {
 
     private static final int NUM_RETRIES = 5;
     public void uploadTsvFile() {
+        if (StringUtils.isEmpty(config.getIuploadURL())) {
+            log.info("Skipping upload because iuploadurl is empty.");
+            return;
+        }
+
         final String iuploadUrl = String.format("%s/%s/file/", config.getIuploadURL(), config.getIndexName());
 
         log.info("Uploading to " + iuploadUrl);
